+++ sdmx-dl-format-protobuf/src/main/java/module-info.java
@@ -4,7 +4,7 @@ module sdmxdl.format.protobuf {
     requires static nbbrd.service;
     requires static org.checkerframework.checker.qual;
 
-    requires transitive sdmxdl.format.util;
+    requires transitive sdmxdl.format.base;
     requires transitive com.google.protobuf;
 
     exports sdmxdl.format.protobuf;