+++ sdmx-dl-provider-base/src/main/java/module-info.java
@@ -1,11 +1,11 @@
-module sdmxdl.provider.util {
+module sdmxdl.provider.base {
 
     requires static lombok;
     requires static nbbrd.design;
     requires static nbbrd.service;
     requires static org.checkerframework.checker.qual;
 
-    requires transitive sdmxdl.format.util;
+    requires transitive sdmxdl.format.base;
 
     exports sdmxdl.provider;
     exports sdmxdl.provider.ext;